Revert "eraer"
This reverts commit 74cb65e8647a633b7a36b9ebb5dd0dbb16379d33.

+++ Assets/Scripts/NonPlayerCharacters/NonPlayerCharacterManager.cs
@@ -464,7 +464,7 @@ namespace VoidRogues.NonPlayerCharacters
             }
 
             // ── RTT-derived effective push-out timeout ────────────────────────────────
-            float playerRtt        = (float)Runner.GetPlayerRtt(Runner.LocalPlayer);
+            float playerRtt        = Runner.GetPlayerRtt(Runner.LocalPlayer);
             float effectiveTimeout = Mathf.Clamp(playerRtt * _rttPushMultiplier, _minPushOutTime, _maxPushOutTime);
 
             // ── Normalised local-player movement direction (used for lateral steering) ─
